+++ package.xml
@@ -2,8 +2,8 @@
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
   <name>hand_gesture_detection</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
+  <version>1.1.6</version>
+  <description>TogetheROS hand gesture detection</description>
   <maintainer email="kao.zhu@horizon.ai">zhukao</maintainer>
   <license>Apache License 2.0</license>
 
@@ -15,6 +15,8 @@
   <depend>rclcpp</depend>
   <depend>dnn_node</depend>
   <depend>ai_msgs</depend>
+
+  <exec_depend>hand_lmk_detection</exec_depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
